Move quantize.py under torchchat/util
